+++ packages/orq-rc/RELEASES.md
@@ -2789,3 +2789,13 @@ Based on:
 - [typescript v3.10.0-rc.37] packages/orq-rc
 ### Releases
 - [NPM v3.10.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.37 - packages/orq-rc
+
+## 2025-08-01 10:59:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.595.1 (2.670.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.10.0-rc.38] packages/orq-rc
+### Releases
+- [NPM v3.10.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.38 - packages/orq-rc